Update no projeto

+++ UniClima/src/java/br/uninove/uniclima/tempo/Main.java
@@ -19,7 +19,7 @@ public class Main {
     private Double tempMin;
     @SerializedName("temp_max")
     @Expose
-    private Integer tempMax;
+    private Double tempMax;
     @SerializedName("pressure")
     @Expose
     private Integer pressure;
@@ -51,11 +51,11 @@ public class Main {
         this.tempMin = tempMin;
     }
 
-    public Integer getTempMax() {
+    public Double getTempMax() {
         return tempMax;
     }
 
-    public void setTempMax(Integer tempMax) {
+    public void setTempMax(Double tempMax) {
         this.tempMax = tempMax;
     }
